Move to a single Vagrantfile
Previous implementation used a bash script to generate a Vagrantfile,
and then call vagrant up.

In this patch, we move away from that behaviour. There is a single
Vagrantfile, and a configuration file (directory.conf.yml) that contains
the configuration of all VMs.

To create a new vm, add it to directory.conf.yml. To remove it (after
running vagrant destroy), remove it from that file.